stredsearch/requirements.txt
attrs>=20.1.0
idna
outcome
sortedcontainers
cryptography
psycopg
psycopg2
django
djangorestframework
markdown
django-filter
django-rq
django-redis
redis
hiredis
django-environ